feat: mobile block component로 분리
